+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## in progress
+
++ Added unit tests
++ Add option to provide message for the prefix.
+
 ## 1.0.0 (21. December 2018)
 
 + First stable release.